+++ hpc/nvidia/cuda/adapter/cufft.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
-/* Copyright (C) 2026 Rong Tao */
+/* Copyright (C) 2025-2026 Rong Tao */
 #ifndef __CUDA_ADAPTER_FFT_H
 #define __CUDA_ADAPTER_FFT_H 1