+++ hpc/nvidia/cuda/adapter/cupti.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
-/* Copyright (C) 2026 Rong Tao */
+/* Copyright (C) 2025-2026 Rong Tao */
 /******************************************************************************\
  *             PTI - CUDA Profiling Tools Interface                           *
 \******************************************************************************/